+++ Sources/DTModelStorage/SectionModel.swift
@@ -50,10 +50,6 @@ open class SectionModel: Section, SectionLocatable
     open var currentSectionIndex: Int? {
         return sectionLocationDelegate?.sectionIndex(for: self)
     }
-
-    @available(*, unavailable, message: "Please use storage.supplementaryModelProvider as a replacement.")
-    /// Supplementaries dictionary.
-    open var supplementaries = [String: [Int: Any]]()
     
     /// Creates section model.
     public init(items: [Any] = []) {